requirements.txt
requests
Flask
python-dotenv
Werkzeug
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
selenium>=4.8.0
requests>=2.28.0
google-auth>=2.20.0
gspread>=5.7.0
oauth2client>=4.1.3

